fix(db): preserve migration boundaries during reset
