+++ turbo/packages/db/scripts/reset-db.ts
@@ -1,10 +1,7 @@
 #!/usr/bin/env tsx
 
-import { drizzle } from "drizzle-orm/postgres-js";
-import { migrate } from "drizzle-orm/postgres-js/migrator";
-import { sql as rawSql } from "drizzle-orm";
 import postgres from "postgres";
-import { DRIZZLE_MIGRATE_OUT } from "../drizzle.config";
+import { applyPendingMigrations } from "./migration-runner";
 
 async function resetDatabase() {
   if (!process.env.DATABASE_URL) {
@@ -12,18 +9,15 @@ async function resetDatabase() {
   }
 
   const sql = postgres(process.env.DATABASE_URL, { max: 1 });
-  const db = drizzle(sql);
 
   try {
     console.log("Dropping all tables...");
-    await db.execute(rawSql`DROP SCHEMA public CASCADE`);
-    await db.execute(rawSql`CREATE SCHEMA public`);
-    await db.execute(rawSql`DROP SCHEMA IF EXISTS drizzle CASCADE`);
+    await sql`DROP SCHEMA public CASCADE`;
+    await sql`CREATE SCHEMA public`;
+    await sql`DROP SCHEMA IF EXISTS drizzle CASCADE`;
 
     console.log("Running migrations...");
-    await migrate(db, {
-      migrationsFolder: DRIZZLE_MIGRATE_OUT,
-    });
+    await applyPendingMigrations(sql);
 
     console.log("Database reset complete");
   } finally {
